# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Nimiq Simple Faucet
  version: 0.0.1
  description: Public, admin, and MCP surfaces for the Nimiq Simple Faucet
    (testnet). See /SECURITY.md for the threat model.
  license:
    name: MIT
    identifier: MIT
  contact:
    name: Security
    url: https://github.com/onmax/Nimiq-Simple-Faucet/blob/main/SECURITY.md
servers:
  - url: /
    description: This instance
...
components:
  securitySchemes:
    integratorHmac:
      type: apiKey
      in: header
      name: X-Faucet-Api-Key
      description: "Integrator HMAC auth: include `X-Faucet-Api-Key`,
        `X-Faucet-Timestamp`, `X-Faucet-Nonce`, and `X-Faucet-Signature`
        (base64-HMAC-SHA256 over
        `method\\npath\\ntimestamp\\nnonce\\nbodyHash`)."
    adminSession:
      type: apiKey
      in: cookie
      name: faucet_session
      description: Session cookie issued by POST /admin/auth/login. Mutating calls
        also require the `X-Faucet-Csrf` double-submit header matching the
        `faucet_csrf` cookie.
    adminMcpToken:
      type: apiKey
      in: header
      name: X-Faucet-Admin-Token
      description: Required to invoke admin-scoped MCP tools over POST /mcp.
  schemas:
    HostContext:
      type: object
      properties:
        uid:
          type: string
          maxLength: 256
        cookieHash:
          type: string
          maxLength: 256
        sessionHash:
          type: string
          maxLength: 256
        accountAgeDays:
          type: integer
          minimum: 0
        emailDomainHash:
          type: string
          maxLength: 256
        kycLevel:
          type: string
          enum:
            - none
            - email
            - phone
            - id
        tags:
          type: array
          items:
            type: string
            maxLength: 64
          maxItems: 32
        signature:
          type: string
          maxLength: 512
    Fingerprint:
      type: object
      properties:
        visitorId:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        components:
          type: object
          additionalProperties: {}
    ClaimRequest:
      type: object
      properties:
        address:
          type: string
          minLength: 1
          description: Target address in native chain format
        captchaToken:
          type: string
        hashcashSolution:
          type: string
        powSolution:
          type: string
          description: Deprecated alias for hashcashSolution
        idempotencyKey:
          type: string
          maxLength: 128
          description: Idempotency key for safe retries
        fingerprint:
          $ref: "#/components/schemas/Fingerprint"
        hostContext:
          $ref: "#/components/schemas/HostContext"
      required:
        - address
    ClaimResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum:
            - broadcast
            - confirmed
            - rejected
            - challenged
            - timeout
            - expired
            - manual-allow
        decision:
          type: string
          enum:
            - allow
            - deny
            - review
            - challenge
        txId:
          type: string
        reason:
          type: string
      required:
        - id
        - status
    ClaimStatus:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        address:
          type: string
        amountLuna:
          type: string
        txId:
          type:
            - string
            - "null"
        createdAt:
          anyOf:
            - type: string
            - type: number
        decision:
          type:
            - string
            - "null"
        rejectionReason:
          type:
            - string
            - "null"
      required:
        - id
        - status
        - address
        - amountLuna
        - txId
        - createdAt
        - decision
        - rejectionReason
...
    HashcashChallenge:
      type: object
      properties:
        challenge:
          type: string
        difficulty:
          type: integer
        expiresAt:
          type: integer
          description: Unix ms timestamp
      required:
        - challenge
        - difficulty
        - expiresAt
    ChallengeRequest:
      type: object
      properties:
        uid:
          type: string
          maxLength: 128
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
        issues:
          type: array
          items: {}
      required:
        - error
...
paths:
...
  /v1/challenge:
    post:
      tags:
        - Public
      summary: Mint a hashcash challenge
      description: Returns a fresh proof-of-work challenge bound to the caller IP.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChallengeRequest"
      responses:
        "200":
          description: Challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HashcashChallenge"
        "404":
          description: Hashcash not enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/claim:
    post:
      tags:
        - Public
      summary: Submit a faucet claim
      description: Evaluates the abuse pipeline and, on `allow`, broadcasts a
        transaction. Optional integrator HMAC headers are verified when
        `X-Faucet-Api-Key` is present.
      security:
        - {}
        - integratorHmac: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimRequest"
      responses:
        "200":
          description: Broadcast
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimResponse"
        "202":
          description: Challenge or review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Integrator auth failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimResponse"
  /v1/claim/{id}:
    get:
      tags:
        - Public
      summary: Claim status by id
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimStatus"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...